♻️🏢 Rename card files
