+++ apps/admin/src/routes/_layout.dashboard/card-active-animals.tsx
