+++ apps/admin/src/routes/_layout.dashboard/card-animals-to-sterilize.tsx
