+++ apps/admin/src/routes/_layout.dashboard/card-dogs-to-diagnose.tsx
