+++ apps/admin/src/routes/_layout.dashboard/card-managed-animals.tsx
