+++ apps/admin/src/routes/_layout.dashboard/route.tsx
@@ -1,11 +1,11 @@
 import { PageLayout } from "#core/layout/page";
 import { getPageTitle } from "#core/page-title";
 import type { MetaFunction } from "@remix-run/node";
-import { ActiveAnimalsCard } from "./active-animals-card";
-import { AnimalsToSterilizeCard } from "./animals-to-sterilize-card";
-import { AnimalsToVaccinateCard } from "./animals-to-vaccinate-card";
-import { DogsToDiagnoseCard } from "./dogs-to-diagnose-card";
-import { ManagedAnimalsCard } from "./managed-animals-card";
+import { ActiveAnimalsCard } from "./card-active-animals";
+import { AnimalsToSterilizeCard } from "./card-animals-to-sterilize";
+import { AnimalsToVaccinateCard } from "./card-animals-to-vaccinate";
+import { DogsToDiagnoseCard } from "./card-dogs-to-diagnose";
+import { ManagedAnimalsCard } from "./card-managed-animals";
 
 export { loader } from "./loader.server";
